// src/modules/service-order/domain/enums/service-order-status.enum.ts
export enum ServiceOrderStatus {
  RECEIVED = 'RECEIVED',
  IN_DIAGNOSIS = 'IN_DIAGNOSIS',
  AWAITING_APPROVAL = 'AWAITING_APPROVAL',
  AWAITING_PARTS = 'AWAITING_PARTS',
  IN_PROGRESS = 'IN_PROGRESS',
  COMPLETED = 'COMPLETED',
  AWAITING_PAYMENT = 'AWAITING_PAYMENT',
  DELIVERED = 'DELIVERED',
  CANCELLED = 'CANCELLED',
}